Fix AbstractCatalogResource (COBJ and CCOL) for version 1.15
Made resource only use ambienceFileInstanceId and isOverrideInstance for exactly auralPropertiesVersion 3 and added a new unknown01 byte for version 4. This seems to work for now. The following properties are in order again now. This may still be completely wrong as the parts we believe to be unused may be simply be used for different purposes now.